Service now caters for unintialised collections on notifications
