+++ .mocharc.json
@@ -1,6 +1,6 @@
 {
 	"extension": ["ts"],
 	"spec": "./tests/**/*.spec.*",
-	"require": ["ts-node/register", "@babel/register", "jsdom-global/register"],
+	"require": ["ts-node/register", "@babel/register"],
 	"exclude": ["node_modules", "./types", "./docs", "./preloads", "./services"]
 }
